Add abstraction of vote
