+++ server/app/states/loup_garou_vote.js
@@ -1,4 +1,5 @@
 const LOUP_GAROU_RESULT = require("./loup_garou_result.js");
+const abstract_vote = require("../abstract_vote.js")
 
 let state = {
   name: "LOUP_GAROU_VOTE",
@@ -9,22 +10,8 @@ let state = {
 }
 
 state.actions.push({
-  name: "loup_garou_vote",
-  fct:  function(game, player, param)  {
-    if (game.state.name === state.name && !(player.vote)) {
-      var candidate = game.players.findByPseudo(param.player_pseudo);
-      if (candidate) {
-        if (candidate.vote_count) {
-          candidate.vote_count++;
-        } else {
-          candidate.vote_count = 1;
-        }
-      }
-      player.vote = candidate;
-      console.info(player.pseudo + " voted for " + param.player_pseudo);
-      game.updateAllPlayers();
-    }
-  }
+  name: "vote",
+  fct: abstract_vote.vote
 });
 
 module.exports = state;
